bump version: 0.4.2-alpha
