+++ tests/unit/components/desktop-panel-test.js
@@ -11,7 +11,9 @@ import {
 } from 'ember-shell/system/panel';
 import { describeComponent, it } from 'ember-mocha';
 
+import ClockDate from 'ember-shell/components/panel/clock-date';
 import MenuButton from 'ember-shell/components/panel/menu-button';
+import StatusMenu from 'ember-shell/components/panel/status-menu';
 import TestItem from 'ember-shell/components/panel/test-item';
 
 const {
@@ -27,25 +29,29 @@ describeComponent(
     let panelManager;
 
     beforeEach(function() {
+      this.register('component:shell/panel/clock-date', ClockDate);
+      this.register('component:shell/panel/status-menu', StatusMenu);
       this.register('component:shell/panel/menu-button', MenuButton);
+      this.register('component:shell/panel/test-item', TestItem);
+
       panelManager = new PanelManager(getOwner(this));
     });
 
     describe('PrimaryPanelInstance', function() {
 
-      let component;
+      let panelComponent;
+      let panelInstance;
 
       beforeEach(function() {
-        let panel = panelManager.panels.get('firstObject');
-        component = this.subject({ panel });
+        panelInstance = panelManager.panels.get('firstObject');
+        panelComponent = this.subject({ panel: panelInstance });
       });
 
       it('should have menu-button item on it', function() {
-        let panel = component.get('panel');
-
-        let hasMenuButton = component.get('items').any((item) =>
+        let hasMenuButton = panelComponent.get('items').any((item) =>
           item.name === "menu-button"
         );
+
         expect(hasMenuButton).to.be.true;
       });
 
@@ -53,29 +59,32 @@ describeComponent(
 
     describe('DefaultPanelInstance', function() {
 
-      let component;
+      let panelComponent;
+      let panelInstance;
 
       beforeEach(function() {
-        let panel = panelManager.addPanel();
-        component = this.subject({ panel });
+        panelInstance = panelManager.addPanel();
+        panelComponent = this.subject({ panel: panelInstance });
       });
 
       it('should have a Panel instance', function() {
-        let panel = component.get('panel');
-        expect(panel).to.exist;
-        expect(Panel.detectInstance(panel)).to.be.true;
+        panelInstance = panelComponent.get('panel');
+
+        expect(panelInstance).to.exist;
+        expect(Panel.detectInstance(panelInstance)).to.be.true;
       });
 
       it('should have a list of panel items', function() {
-        let items = component.get('items');
+        let items = panelComponent.get('items');
+
         expect(items).to.exist;
         expect(Ember.isArray(items)).to.be.true;
       });
 
       it('should allow to insert a panel item into a panel', function() {
-        let panelItem = panelManager.insertItem(panel, 'test-item');
+        let panelItem = panelManager.insertItem(panelInstance, 'test-item');
 
-        expect(panel.items.contains(panelItem)).to.be.true;
+        expect(panelInstance.items.contains(panelItem)).to.be.true;
       });
 
     });
