+++ tests/unit/mixins/behavior/focusable-test.js
@@ -2,13 +2,13 @@
 import { expect } from 'chai';
 import { describe, it } from 'mocha';
 import Ember from 'ember';
-import PressableMixin from 'ember-shell/mixins/behaviour/pressable';
+import FocusableMixin from 'ember-shell/mixins/behavior/focusable';
 
-describe('UiPressableMixin', function() {
+describe('FocusableMixin', function() {
   // Replace this with your real tests.
   it('works', function() {
-    let UiPressableObject = Ember.Object.extend(PressableMixin);
-    let subject = UiPressableObject.create();
+    let UiFocusableObject = Ember.Object.extend(FocusableMixin);
+    let subject = UiFocusableObject.create();
     expect(subject).to.be.ok;
   });
 });
