+++ tests/unit/mixins/behavior/scalable-test.js
@@ -2,7 +2,7 @@
 import { expect } from 'chai';
 import {describe, it } from 'mocha';
 import Ember from 'ember';
-import ScalableMixin from 'ember-shell/mixins/behaviour/scalable';
+import ScalableMixin from 'ember-shell/mixins/behavior/scalable';
 
 describe('UiScalableMixin', function() {
   // Replace this with your real tests.
